Add files via upload
What is Manual Testing?
Manual Testing is a kind of software testing in which a software tester develops and executes the test cases without using any automated testing tools. The main objective of manual testing is to detect the issues, bugs, and defects of a software application. Any new software application should be manually tested before performing the automation testing. The software testing fundamental “100% Automation is not possible” makes Manual Testing essential.

Manual Testing Concepts
Need of Manual Testing
Whenever a software application is released into the market, it may have some issues or bugs, or end-users may face some problems while using it. So, we have to perform one testing stage to make the application stable and bug-free and deliver a quality software product to users. While performing the manual testing, the test engineer tests the software from the customer’s perspective to develop the appropriate test cases and give proper feedback to the developers.

Testing Methods
We have two types of testing methods, they are:

#1. Static Testing
Static Testing is also known as Verification. It is a method of verifying the files and documents. It ensures that we are developing the correct product. It also verifies the requirements we have. In this method, test engineers will carry out activities like Reviews, Inspections, Walkthroughs, etc.

#2. Dynamic Testing
Dynamic Testing is also called Validation. It is a dynamic process of testing the software product. It validates whether the developed product is right or not.

How to do Manual Testing?
Follow these steps to do manual testing:

Step1: First, review all the documents related to software, for selecting the testing areas.

Step2: Analyse all the required documents to get all the requirements mentioned by the end-user.

Step3: Build test cases as per the requirement document.

Step4: Execute all the test cases manually by using white-box testing and black-box testing.

Step5: If bugs are detected, report them to the development team.

Step6: When the developer fixes the bug, retest it.